Fix errors for success records
